backend/requirements.txt
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
requests==2.26.0
python-dotenv
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
Pillow==9.5.0
django-extra-fields==3.0.2
djoser